Cache Gxact for 2pc.
This is because on greeenplum mostly commit/abort prepared should be usually on
the same backend of prepare and we can accelerate them without searching the
whole prepared gxact array. array->hash table might be another solution but
seems does not desevere so much effort also hash table operation has overhead
also.

upstream also cached_gxact and cached_xid but that's for recovery and seems to
be not that usefaul since in heavy-load environment. 2pc xlog with various gids
should be interleaved so maybe propose upstream to remove them?